+++ apps/backoffice/src/lib/be/services/business.ts
@@ -4,16 +4,8 @@ import {
 } from "@/config/constants";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 
-import { ServiceLifecycleStatus } from "@/generated/api/ServiceLifecycleStatus";
-import { ServiceLifecycleStatusTypeEnum } from "@/generated/api/ServiceLifecycleStatusType";
 import { ServiceList } from "@/generated/api/ServiceList";
-import {
-  ServiceListItem,
-  VisibilityEnum
-} from "@/generated/api/ServiceListItem";
-import { CategoryEnum, ScopeEnum } from "@/generated/api/ServiceMetadata";
 import { getApimRestClient } from "@/lib/be/apim-service";
-import { ServiceLifecycle, ServicePublication } from "@io-services-cms/models";
 import { readableReport } from "@pagopa/ts-commons/lib/reporters";
 import * as E from "fp-ts/lib/Either";
 import * as RA from "fp-ts/lib/ReadonlyArray";
@@ -26,92 +18,13 @@ import {
   retrieveLifecycleServices,
   retrievePublicationServices
 } from "./cosmos";
+import { reducePublicationServicesList, toServiceListItem } from "./utils";
 
 type PathParameters = {
   serviceId?: string;
   keyType?: string;
   limit?: string;
   offset?: string;
-};
-
-const reducePublicationServicesList = (
-  publicationServices: ReadonlyArray<ServicePublication.ItemType>
-) =>
-  pipe(
-    publicationServices,
-    RA.map(item => [item.id, item.fsm.state] as [string, VisibilityEnum]),
-    arr =>
-      arr.reduce((acc, [key, value]) => {
-        acc[key] = value;
-        return acc;
-      }, {} as Record<string, VisibilityEnum>)
-  );
-
-export const toServiceListItem = ({
-  fsm,
-  data,
-  id,
-  last_update
-}: ServiceLifecycle.ItemType): ServiceListItem => ({
-  id,
-  status: toServiceStatus(fsm),
-  last_update: last_update ?? new Date().getTime().toString(),
-  name: data.name,
-  description: data.description,
-  organization: data.organization,
-  metadata: {
-    ...data.metadata,
-    scope: toScopeType(data.metadata.scope),
-    category: toCategoryType(data.metadata.category)
-  },
-  authorized_recipients: data.authorized_recipients,
-  authorized_cidrs: data.authorized_cidrs
-});
-
-const toServiceStatus = (
-  fsm: ServiceLifecycle.ItemType["fsm"]
-): ServiceLifecycleStatus => {
-  switch (fsm.state) {
-    case "approved":
-    case "deleted":
-    case "draft":
-    case "submitted":
-      return { value: ServiceLifecycleStatusTypeEnum[fsm.state] };
-    case "rejected":
-      return {
-        value: ServiceLifecycleStatusTypeEnum[fsm.state],
-        reason: (fsm.reason as string) ?? undefined // FIXME
-      };
-
-    default:
-      const _: never = fsm;
-      return ServiceLifecycleStatusTypeEnum[fsm];
-  }
-};
-
-const toScopeType = (
-  s: ServiceLifecycle.ItemType["data"]["metadata"]["scope"]
-): ScopeEnum => {
-  switch (s) {
-    case "LOCAL":
-    case "NATIONAL":
-      return ScopeEnum[s];
-    default:
-      const _: never = s;
-      return ScopeEnum[s];
-  }
-};
-
-const toCategoryType = (
-  s: ServiceLifecycle.ItemType["data"]["metadata"]["category"]
-): CategoryEnum => {
-  switch (s) {
-    case "STANDARD":
-    case "SPECIAL":
-      return CategoryEnum[s];
-    default:
-      return CategoryEnum.STANDARD;
-  }
 };
 
 /**
